Add remote_write Appender to promxy
This makes promxy use the `remote_write` config in the prometheus
config. This is used to configure the local appender which is used for
metrics generated by prom itself (alerting and recording rules).

Fixes #79
