+++ cmd/promxy/main.go
@@ -244,7 +244,27 @@ func main() {
 			}
 			files = append(files, fs...)
 		}
-		return ruleManager.Update(time.Duration(cfg.GlobalConfig.EvaluationInterval), files)
+		if err := ruleManager.Update(time.Duration(cfg.GlobalConfig.EvaluationInterval), files); err != nil {
+			return err
+		}
+		return nil
+
+		if cfg.RemoteWriteConfigs == nil {
+			ruleList := ruleManager.Rules()
+			// check for any recording rules, if we find any lets log a fatal and stop
+			// https://github.com/jacksontj/promxy/issues/74
+			for _, rule := range ruleList {
+				if _, ok := rule.(*rules.RecordingRule); ok {
+					return fmt.Errorf("Promxy doesn't support recording rules: %s", rule)
+				}
+			}
+
+			if len(ruleList) > 0 {
+				logrus.Warning("Alerting rules are configured but no remote_write endpoint is configured.")
+			}
+		}
+
+		return nil
 	}}))
 
 	// We need an empty scrape manager, simply to make the API not panic and error out
@@ -314,22 +334,6 @@ func main() {
 		logrus.Fatalf("Error loading config: %s", err)
 	}
 
-	// Our own checks on top of the config
-	checkConfig := func() error {
-		// check for any recording rules, if we find any lets log a fatal and stop
-		// https://github.com/jacksontj/promxy/issues/74
-		for _, rule := range ruleManager.Rules() {
-			if _, ok := rule.(*rules.RecordingRule); ok {
-				return fmt.Errorf("Promxy doesn't support recording rules: %s", rule)
-			}
-		}
-		return nil
-	}
-
-	if err := checkConfig(); err != nil {
-		logrus.Fatalf("Error checking config: %v", err)
-	}
-
 	close(reloadReady)
 
 	// Set up access logger
@@ -372,9 +376,6 @@ func main() {
 				log.Infof("Reloading config")
 				if err := reloadConfig(reloadables...); err != nil {
 					log.Errorf("Error reloading config: %s", err)
-				}
-				if err := checkConfig(); err != nil {
-					logrus.Errorf("Error checking config: %v", err)
 				}
 			case syscall.SIGTERM, syscall.SIGINT:
 				log.Infof("promxy exiting")
